Added some tests

+++ src/test/java/PositionTest.java
@@ -1,4 +1,5 @@
-import com.aor.soulsduel.model.Position;
+
+import com.aor.hero.model.Position;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -14,8 +15,8 @@ public class PositionTest {
         }
         @Test
         public void setPositionTest(){
-            int xp= position.getX();
-            int yp = position.getY();
+            int xp= (int) position.getX();
+            int yp = (int) position.getY();
             Assertions.assertEquals(10,xp);
             Assertions.assertEquals(5,yp);
         }
